GameManager.cs: 데이터타입을 v2에 맞춰 수정

+++ Assets/Scripts/Server/GameManager.cs
@@ -23,7 +23,7 @@ public class GameManager : MonoBehaviorSingleton<GameManager>
         if (packet.Success)
         {
             Debug.Log($"✓ Login Success! Player ID: {packet.Player.Id}, Name: {packet.Player.Name}");
-            NetManager.Instance._playerId = (int)packet.Player.Id; //  이런 캐스팅 부분 나중에 수정해야함.
+            NetManager.Instance._playerId = packet.Player.Id;
             NetManager.Instance.PlayerName = packet.Player.Name;
             NetManager.Instance.PlayerTag = packet.Player.Tag;
             // 로그인 성공 → Splash → Main → (멀티플레이 버튼 시 방 생성 후 로비)
